+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Switch to `/` is not posible because this path was considered invalid
 - Fixed color scheme with white terminals
+- Fixed hang when switching to invalid path
 
 
 ## [0.1.0] - 2023-11-10